+++ src/user/user_api.cc
@@ -2250,13 +2250,13 @@ int mjs_setName(mjsElement* element, const char* name) {
     return 0;
   }
   mjCBase* baseC = static_cast<mjCBase*>(element);
-  baseC->name    = std::string(name);
   try {
-    baseC->model->CheckRepeat(element->elemtype);
+    baseC->model->CheckNameChange(element->elemtype, baseC->name, name);
   } catch (mjCError& e) {
     baseC->model->SetError(e);
     return -1;
   }
+  baseC->name = std::string(name);
   return 0;
 }