+++ src/user/user_objects.cc
@@ -1965,8 +1965,8 @@ mjCBody* mjCBody::AddBody(mjCDef* _def) {
 
   obj->parent = this;
 
-  // update signature
-  model->spec.element->signature = model->Signature();
+  // invalidate signature
+  model->InvalidateSignature();
   return obj;
 }
 
@@ -1978,8 +1978,8 @@ mjCFrame* mjCBody::AddFrame(mjCFrame* _frame) {
   model->ResetTreeLists();
   model->MakeTreeLists();
 
-  // update signature
-  model->spec.element->signature = model->Signature();
+  // invalidate signature
+  model->InvalidateSignature();
   return obj;
 }
 
@@ -1999,8 +1999,8 @@ mjCJoint* mjCBody::AddFreeJoint() {
   model->ResetTreeLists();
   model->MakeTreeLists();
 
-  // update signature
-  model->spec.element->signature = model->Signature();
+  // invalidate signature
+  model->InvalidateSignature();
   return obj;
 }
 
@@ -2019,8 +2019,8 @@ mjCJoint* mjCBody::AddJoint(mjCDef* _def) {
   model->ResetTreeLists();
   model->MakeTreeLists();
 
-  // update signature
-  model->spec.element->signature = model->Signature();
+  // invalidate signature
+  model->InvalidateSignature();
   return obj;
 }
 
@@ -2039,8 +2039,8 @@ mjCGeom* mjCBody::AddGeom(mjCDef* _def) {
   model->ResetTreeLists();
   model->MakeTreeLists();
 
-  // update signature
-  model->spec.element->signature = model->Signature();
+  // invalidate signature
+  model->InvalidateSignature();
   return obj;
 }
 
@@ -2059,8 +2059,8 @@ mjCSite* mjCBody::AddSite(mjCDef* _def) {
   model->ResetTreeLists();
   model->MakeTreeLists();
 
-  // update signature
-  model->spec.element->signature = model->Signature();
+  // invalidate signature
+  model->InvalidateSignature();
   return obj;
 }
 
@@ -2079,8 +2079,8 @@ mjCCamera* mjCBody::AddCamera(mjCDef* _def) {
   model->ResetTreeLists();
   model->MakeTreeLists();
 
-  // update signature
-  model->spec.element->signature = model->Signature();
+  // invalidate signature
+  model->InvalidateSignature();
   return obj;
 }
 
@@ -2099,8 +2099,8 @@ mjCLight* mjCBody::AddLight(mjCDef* _def) {
   model->ResetTreeLists();
   model->MakeTreeLists();
 
-  // update signature
-  model->spec.element->signature = model->Signature();
+  // invalidate signature
+  model->InvalidateSignature();
   return obj;
 }
 
@@ -2132,7 +2132,7 @@ mjCFrame* mjCBody::ToFrame() {
                        parent->bodies.end());
   model->ResetTreeLists();
   model->MakeTreeLists();
-  model->spec.element->signature = model->Signature();
+  model->InvalidateSignature();
   return newframe;
 }